requirements.txt
# --- Core API ---
Flask==3.0.2
flask-restx==1.3.0
flask-cors==4.0.0

# --- Database & Auth ---
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-jwt-extended==4.6.0

# --- NLP ---
spacy==3.7.4

# --- Development & Testing ---
bandit==1.7.7
pytest==8.0.2
python-dotenv==1.0.1